backend/pyproject.toml
[project]
name = "neurovet-backend"
version = "1.0.0"
description = "Veterinary Neurological Diagnostic Assistant Backend"
authors = [
    {name = "NeuroVet Team", email = "contact@neurovet.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.7.0",
    "openai>=1.50.0",
    "sqlalchemy>=2.0.36",
    "aiomysql>=0.2.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "alembic>=1.16.5",
    "gunicorn>=23.0.0",
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "email-validator>=2.3.0",
    "resend>=2.19.0",
]

[project.scripts]
neurovet-backend = "main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
